server/requirements.txt
alembic==1.14.0
asttokens==2.4.1
blinker==1.9.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
decorator==5.1.1
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.2.2
executing==2.1.0
Faker==33.1.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
httplib2==0.20.2
idna==2.8
importlib-metadata==4.6.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
matplotlib-inline==0.1.7
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pipenv==2024.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==8.3.3
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2024.2
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
stack-data==0.6.3
systemd-python==234
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.25.11
virtualenv==20.26.6
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==1.0.0
